refactor: 최단경로와 총 거리 구하는 기능 분리
