+++ docs/README.md
@@ -65,10 +65,9 @@
 
 - 요금 기능
   - [ ] 기본 운임(10km)은 1,250원이다.
-  - [ ] 10km ~ 50km로 운인 거리를 초과했을 경우 5km 마다 100원 추가
-  - [ ] 50km 이상 운인 거리를 초과했을 경우 8km 마다 100원 추가
+  - [ ] 10km ~ 50km로 운임 거리를 초과했을 경우 5km 마다 100원 추가
+  - [ ] 50km 이상 운임 거리를 초과했을 경우 8km 마다 100원 추가
   
-
 ---
 
 ## 프로그래밍 요구 사항
